chore: 调整了Demo/main_page_demo下的目录结构
