--- Demo/main_page_demo/css/components/game_panels.css
@@ -1,175 +1,0 @@
-/* ===== 游戏面板和存档面板样式 ===== */
-
-/* 通用面板头部样式 */
-.panel-header {
-    display: flex;
-    align-items: center;
-    padding: 15px 20px;
-    position: relative;
-    justify-content: space-between;
-    border-bottom: 1px solid rgba(255, 255, 255, 0.1);
-}
-
-.panel-header h2 {
-    margin: 0;
-    color: white;
-    font-size: 24px;
-    font-weight: 600;
-    text-shadow: 0 2px 4px rgba(0, 0, 0, 0.3);
-}
-
-/* 通用关闭按钮样式 */
-.panel-header .close-button {
-    background: transparent;
-    border: none;
-    padding: 8px;
-    border-radius: 50%;
-    cursor: pointer;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    color: white;
-    transition: all 0.3s ease-in-out;
-    font-size: 24px;
-    width: 40px;
-    height: 40px;
-}
-
-.panel-header .close-button:hover {
-    color: var(--accent-color);
-    background-color: rgba(255, 255, 255, 0.1);
-    transform: rotate(90deg);
-}
-
-/* 通用面板内容区域 */
-.panel-body {
-    flex: 1;
-    padding: 20px;
-    overflow-y: auto;
-    display: flex;
-    flex-direction: column;
-}
-
-/* ===== 游戏面板特定样式 ===== */
-.game-screen-panel {
-    display: flex;
-    flex-direction: column;
-}
-
-.game-screen-content {
-    background: rgba(255, 255, 255, 0.1);
-    border-radius: 12px;
-    padding: 30px;
-    text-align: center;
-    backdrop-filter: blur(10px);
-    border: 1px solid rgba(255, 255, 255, 0.1);
-}
-
-.game-screen-content h3 {
-    color: var(--accent-color);
-    font-size: 28px;
-    margin-bottom: 20px;
-    text-shadow: 0 2px 4px rgba(0, 0, 0, 0.3);
-}
-
-.game-screen-content p {
-    font-size: 18px;
-    line-height: 1.6;
-    margin-bottom: 30px;
-    color: rgba(255, 255, 255, 0.9);
-}
-
-.game-action-buttons {
-    display: flex;
-    gap: 15px;
-    justify-content: center;
-    flex-wrap: wrap;
-}
-
-.game-action-btn {
-    background: linear-gradient(135deg, var(--accent-color), #64b5f6);
-    color: white;
-    border: none;
-    padding: 12px 24px;
-    border-radius: 25px;
-    font-size: 16px;
-    font-weight: 600;
-    cursor: pointer;
-    transition: all 0.3s ease;
-    box-shadow: 0 4px 15px rgba(121, 217, 255, 0.3);
-}
-
-.game-action-btn:hover {
-    transform: translateY(-2px);
-    box-shadow: 0 6px 20px rgba(121, 217, 255, 0.4);
-}
-
-/* ===== 存档面板特定样式 ===== */
-.load-save-panel {
-    display: flex;
-    flex-direction: column;
-}
-
-.save-files-container {
-    background: rgba(255, 255, 255, 0.1);
-    border-radius: 12px;
-    padding: 20px;
-    backdrop-filter: blur(10px);
-    border: 1px solid rgba(255, 255, 255, 0.1);
-}
-
-.save-files-container h3 {
-    color: var(--accent-color);
-    font-size: 24px;
-    margin-bottom: 20px;
-    text-shadow: 0 2px 4px rgba(0, 0, 0, 0.3);
-}
-
-.save-file-list {
-    display: grid;
-    gap: 12px;
-    max-height: 400px;
-    overflow-y: auto;
-}
-
-.save-file-item {
-    background: rgba(255, 255, 255, 0.05);
-    border: 1px solid rgba(255, 255, 255, 0.1);
-    border-radius: 8px;
-    padding: 15px;
-    cursor: pointer;
-    transition: all 0.3s ease;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-}
-
-.save-file-item:hover {
-    background: rgba(255, 255, 255, 0.1);
-    border-color: var(--accent-color);
-    transform: translateX(5px);
-}
-
-.save-file-info h4 {
-    margin: 0 0 5px 0;
-    color: white;
-    font-size: 16px;
-}
-
-.save-file-info p {
-    margin: 0;
-    color: rgba(255, 255, 255, 0.7);
-    font-size: 14px;
-}
-
-.save-file-date {
-    color: rgba(255, 255, 255, 0.6);
-    font-size: 12px;
-}
-
-.no-saves-message {
-    text-align: center;
-    color: rgba(255, 255, 255, 0.7);
-    font-style: italic;
-    padding: 40px;
-}
