--- Demo/main_page_demo/css/components/main_menu.css
@@ -1,83 +1,0 @@
-/* 主容器，用于设置背景和布局 */
-.main-container {
-    width: 100%;
-    height: 100%;
-    
-    /* 
-     *  关键修改：移除这里的背景。
-     *  背景将完全由 ::before 伪元素来控制。
-     */
-    /*
-    background-image: url('../../../../frontend/public/pictures/backgrounds/login/X0_6rTZl.png');
-    background-size: cover;
-    background-position: center;
-    */
-    
-    display: flex;
-    justify-content: flex-start; /* 将主菜单推到左边 */
-    align-items: center;
-    position: relative; /* 必须保留，为 ::before 提供定位锚点 */
-}
-
-/* 开始页面容器 - 继承主容器的布局属性 */
-#start-page {
-    width: 100%;
-    height: 100%;
-    display: flex;
-    justify-content: flex-start;
-    align-items: center;
-    position: relative;
-}
-
-/* 主菜单 */
-.main-menu {
-    display: flex;
-    flex-direction: column;
-    /* align-items: center; */
-    padding: 20px;
-    margin-left: 10vw; /* 距离左侧边缘10%视口宽度 */
-    position: relative; /* 确保菜单显示在背景之上 */
-    z-index: 1;
-}
-
-.logo {
-    position: absolute;
-    top: 5vh; /* 距离顶部5%视口高度 */
-    right: 5vw; /* 距离右侧5%视口宽度 */
-    height: 40vh; /* 高度为视口高度的40% */
-    width: auto; /* 宽度自动，保持比例 */
-    max-width: 40vw; /* 最大宽度不超过视口宽度的40%，防止过宽 */
-    filter: drop-shadow(0 4px 8px rgba(0,0,0,0.3)); /* 添加一点阴影使其更突出 */
-    z-index: 1; /* 确保logo显示在背景之上 */
-}
-
-.menu-options {
-    display: flex;
-    flex-direction: column;
-    align-items: stretch;
-    width: 350px; /* 按钮加宽一点 */
-}
-
-.menu-button {
-    background: transparent; /* 去除背景 */
-    color: white;
-    border: none; /* 去除边框 */
-    padding: 15px;
-    margin: 10px 0;
-    border-radius: 12px;
-    /* 使用clamp()实现响应式字体大小 */
-    /* 最小32px, 根据视口宽度的4%缩放, 最大72px */
-    font-size: clamp(32px, 4vw, 72px);
-    font-weight: normal; /* 字体加粗 */
-    font-family: 'Maoken Assorted Sans', -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, "Helvetica Neue", Arial, sans-serif; /* 应用自定义字体，并提供备用字体 */
-    cursor: pointer;
-    transition: color 0.3s, text-shadow 0.3s; /* 平滑过渡 */
-    text-shadow: 0 2px 4px rgba(0,0,0,0.5); /* 加一点文字阴影以保证清晰度 */
-    text-align: left; /* 文字左对齐 */
-}
-
-.menu-button:hover {
-    color: #f0f0f0;
-    text-shadow: 0 0 10px rgba(255, 255, 255, 0.8); /* 悬停时发光效果 */
-    transform: none; /* 移除之前的缩放效果 */
-}
