--- Demo/main_page_demo/css/components/page_text.css
@@ -1,173 +1,0 @@
-.action-button.full-width {
-    width: 100%;
-}
-
-/* --- 文本设置页面新样式 --- */
-.settings-columns {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 0px;
-    height: 100%;
-}
-
-.setting-item {
-    margin-bottom: 25px;
-    width: 100%;
-    max-width: 900px;
-}
-
-/* 滑块样式 (移植自 menu.css) */
-.slider-container {
-    display: flex;
-    align-items: center;
-    gap: 15px;
-    color: rgba(255, 255, 255, 0.9); /* 白色文字 */
-    text-shadow: 0 2px 4px rgba(0, 0, 0, 0.3);
-}
-
-input[type="range"] {
-    -webkit-appearance: none;
-    appearance: none;
-    flex-grow: 1;
-    margin: 10px 0;
-    background-color: transparent;
-    position: relative;
-    outline: none;
-}
-
-input[type="range"]::-webkit-slider-runnable-track {
-    width: 100%;
-    height: 8px;
-    border-radius: 4px;
-    background: rgba(255, 255, 255, 0.2); /* 透明白色轨道 */
-    border: 1px solid rgba(255, 255, 255, 0.1);
-    box-shadow: inset 0 1px 2px rgba(0, 0, 0, 0.2);
-    transition: all 0.3s ease;
-}
-
-input[type="range"]:hover::-webkit-slider-runnable-track {
-    height: 10px;
-    background: rgba(255, 255, 255, 0.25);
-}
-
-input[type="range"]::-webkit-slider-thumb {
-    -webkit-appearance: none;
-    appearance: none;
-    width: 22px;
-    height: 22px;
-    background: linear-gradient(135deg, var(--accent-color), #64b5f6); /* 渐变背景，不再是白色 */
-    border-radius: 50%;
-    cursor: grab;
-    border: 2px solid rgba(255, 255, 255, 0.8); /* 半透明白色边框 */
-    margin-top: -7px;
-    transition: all 0.2s cubic-bezier(0.25, 0.8, 0.25, 1);
-    transform-origin: center;
-    box-shadow: 0 4px 12px rgba(121, 217, 255, 0.4), 0 2px 4px rgba(0, 0, 0, 0.2); /* 增强阴影效果 */
-    position: relative;
-    z-index: 2;
-}
-
-input[type="range"]::-webkit-slider-thumb:hover,
-input[type="range"]::-webkit-slider-thumb:active {
-    transform: scale(1.15);
-    box-shadow: 0 6px 20px rgba(121, 217, 255, 0.6), 0 4px 8px rgba(0, 0, 0, 0.3);
-    background: linear-gradient(135deg, #64b5f6, var(--accent-color)); /* 悬停时反转渐变 */
-}
-
-input[type="range"]:active::-webkit-slider-thumb {
-    cursor: grabbing;
-}
-
-/* 开关样式 */
-.toggle-container {
-    display: flex;
-    align-items: center;
-}
-
-.toggle-input {
-    display: none;
-}
-
-.toggle-label {
-    cursor: pointer;
-    position: relative;
-    padding-left: 60px;
-    font-size: 14px;
-    color: white; /* 修改文字颜色为白色 */
-    text-shadow: 0 2px 4px rgba(0, 0, 0, 0.3);
-}
-
-.toggle-label::before {
-    content: '';
-    position: absolute;
-    left: 0;
-    top: 50%;
-    transform: translateY(-50%);
-    width: 50px;
-    height: 26px;
-    background-color: rgba(255, 255, 255, 0.2); /* 透明背景 */
-    border: 1px solid rgba(255, 255, 255, 0.3);
-    border-radius: 13px;
-    transition: all 0.3s ease;
-    backdrop-filter: blur(5px);
-}
-
-.toggle-label::after {
-    content: '';
-    position: absolute;
-    left: 4px;
-    top: 50%;
-    transform: translateY(-50%);
-    width: 20px;
-    height: 20px;
-    background: linear-gradient(135deg, #ffffff, #f0f0f0); /* 轻微渐变的白色 */
-    border-radius: 50%;
-    box-shadow: 0 2px 6px rgba(0, 0, 0, 0.3), 0 1px 2px rgba(0, 0, 0, 0.1);
-    transition: all 0.3s ease;
-}
-
-.toggle-input:checked + .toggle-label::before {
-    background-color: rgba(121, 217, 255, 0.3); /* 半透明主题色 */
-    border-color: var(--accent-color);
-    box-shadow: 0 0 10px rgba(121, 217, 255, 0.3);
-}
-
-.toggle-input:checked + .toggle-label::after {
-    left: 26px;
-    background: linear-gradient(135deg, var(--accent-color), #64b5f6); /* 激活时使用主题色渐变 */
-    box-shadow: 0 3px 8px rgba(121, 217, 255, 0.4), 0 1px 3px rgba(0, 0, 0, 0.2);
-}
-
-/* 文本样本显示区域 */
-.text-sample-display {
-    background: rgba(0, 0, 0, 0.3); /* 半透明黑色背景 */
-    color: #ffffff;
-    padding: 15px 20px;
-    border-radius: 12px;
-    border: 1px solid rgba(255, 255, 255, 0.1);
-    font-family: 'Courier New', Courier, monospace;
-    min-height: 2.5em; /* 确保有足够高度容纳光标 */
-    backdrop-filter: blur(10px);
-    box-shadow: inset 0 2px 8px rgba(0, 0, 0, 0.2);
-}
-
-/* 打字光标动画 */
-.typing-cursor {
-    display: inline-block;
-    height: 1.2em;
-    width: 3px;
-    vertical-align: text-bottom;
-    margin-left: 4px;
-    background-color: var(--accent-color);
-    animation: cursor-blink 0.8s infinite;
-}
-
-@keyframes cursor-blink {
-    0%, 100% {
-        opacity: 1;
-    }
-    50% {
-        opacity: 0;
-    }
-}
