+++ Demo/main_page_demo/v0.0.1_by_yukito/css/style.css
@@ -1,0 +1,21 @@
+/* 定义自定义字体 */
+@font-face {
+    font-family: 'Maoken Assorted Sans';
+    src: url('../../resource/fonts/MaokenAssortedSans.ttf') format('truetype');
+    font-weight: normal;
+    font-style: normal;
+}
+
+:root {
+    --accent-color: #79d9ff; /* 从Logo提取的主题粉色 */
+}
+
+/* 全局样式和字体 */
+body, html {
+    margin: 0;
+    padding: 0;
+    width: 100%;
+    height: 100%;
+    font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, "Helvetica Neue", Arial, sans-serif;
+    overflow: hidden;
+}
