+++ Demo/main_page_demo/v0.0.2_by_风雪/css/components/settings_panel.css
@@ -1,0 +1,172 @@
+/* --- 设置面板新样式 (v2) --- */
+.settings-panel {
+    position: fixed;
+    top: 0;
+    right: 0;
+    width: 100%;
+    height: 100%;
+    transform: translateX(100%);
+    transition: transform 0.6s cubic-bezier(0.8, 0, 0.2, 1), opacity 0.6s cubic-bezier(0.8, 0, 0.2, 1);
+    opacity: 0;
+    padding: 0;
+    box-sizing: border-box;
+    z-index: 1000;
+    color: #333;
+    /* 
+     * 修改点：移除这里的背景色，或设置为 transparent。
+     * 这使得面板本身是透明的，可以透出下方的模糊层。
+     */
+    /* background-color: rgba(0, 0, 0, 0.25); */
+    background-color: transparent;
+}
+
+.settings-header {
+    display: flex;
+    align-items: center;
+    padding: 10px 15px;
+    position: relative;
+    justify-content: space-between;
+    /* 移除背景，让它透明以显示下方的模糊效果 */
+    /* background: rgba(0, 0, 0, 0.2); */
+}
+
+/* 调整设置面板内的文字颜色，使其在透明背景上更清晰 */
+.settings-header h2,
+.settings-nav .nav-button,
+.settings-content {
+    color: white;
+    text-shadow: 0 2px 4px rgba(0, 0, 0, 0.2);
+}
+
+.settings-logo {
+    width: 80px;
+    height: auto;
+    margin-right: 0;
+    /* 添加阴影使logo在透明背景上更显眼 */
+    filter: drop-shadow(0 2px 4px rgba(0, 0, 0, 0.2));
+}
+
+.close-button {
+    background: transparent;
+    border: none;
+    padding: 6px;
+    border-radius: 50%;
+    cursor: pointer;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    color: white; /* 改为白色，更适合透明背景 */
+    transition: all 0.3s ease-in-out;
+}
+
+.close-button:hover {
+    color: var(--accent-color);
+    background-color: rgba(255, 255, 255, 0.1);
+    transform: rotate(90deg);
+}
+
+.close-button svg {
+    width: 28px;
+    height: 28px;
+}
+
+.settings-body {
+    display: flex;
+    height: calc(100% - 68px);
+    margin-top: 0;
+    /* 移除背景，让它透明以显示下方的模糊效果 */
+    /* background: rgba(0, 0, 0, 0.1); */
+}
+
+.settings-nav {
+    position: absolute;
+    left: 50%;
+    transform: translateX(-50%);
+    display: flex;
+    flex-direction: row;
+    padding: 0;
+    height: 100%;
+    align-items: center;
+}
+
+/* 滑动指示器样式调整 */
+.nav-indicator {
+    position: absolute;
+    bottom: 0;
+    left: 0;
+    width: 0;
+    height: 4px;
+    background-color: var(--accent-color);
+    border-radius: 2px;
+    z-index: 1;
+    transition: left 0.3s ease-in-out, width 0.3s ease-in-out;
+    /* 添加发光效果 */
+    box-shadow: 0 0 10px rgba(121, 217, 255, 0.4);
+}
+
+.settings-nav .nav-button {
+    background: none;
+    color: white; /* 改为白色 */
+    text-align: center;
+    padding: 10px 15px;
+    border-radius: 8px;
+    border: none;
+    cursor: pointer;
+    margin: 0 5px;
+    font-size: 16px;
+    font-weight: bold;
+    position: relative;
+    z-index: 1;
+    transition: color 0.3s ease, background-color 0.3s ease;
+    display: flex;
+    align-items: center;
+    gap: 8px;
+}
+
+.settings-nav .nav-button svg {
+    width: 18px;
+    height: 18px;
+    stroke-width: 2.5px;
+    flex-shrink: 0;
+}
+
+.settings-nav .nav-button:hover {
+    color: var(--accent-color);
+}
+
+.settings-nav .nav-button.active {
+    color: var(--accent-color);
+    background-color: rgba(255, 255, 255, 0.1); /* 半透明白色背景 */
+}
+
+.settings-nav .nav-button.active:hover {
+    color: var(--accent-color);
+    background-color: rgba(255, 255, 255, 0.15);
+}
+
+.settings-content {
+    flex-grow: 1;
+    background: none;
+    border-radius: 0;
+    padding: 20px;
+}
+
+.settings-page {
+    display: none;
+    color: white; /* 改为白色 */
+    height: 100%;
+    /* 添加文字阴影提高可读性 */
+    text-shadow: 0 2px 4px rgba(0, 0, 0, 0.2);
+}
+
+.settings-page.active {
+    display: block;
+}
+
+/* 为设置内容添加半透明背景，提高可读性 */
+.settings-page > * {
+    background: rgba(255, 255, 255, 0.1);
+    border-radius: 12px;
+    padding: 15px;
+    margin-bottom: 15px;
+}
