--- Demo/main_page_demo/旧代码/主界面demo.md
@@ -1,54 +1,0 @@
-# 主界面和设置界面 UI Demo 说明
-
-## 1. 关于此 Demo
-
-这是一个纯前端 UI Demo，可独立运行测试，与本项目其他文件互不干扰，旨在为游戏主界面设计和设置面板重构提供**视觉风格**与**交互设计**参考。
-
-## 2. 实现功能概览
-
-此 Demo 模拟了从游戏启动到进行详细设置的整个流程，包含以下主要功能模块：
-
-* **加载动画（由喵哒子大人提供）**：
-
-  * 一个可爱的猫耳主题加载动画，带有平滑的进度条和过渡效果。
-  * 动画结束后自动淡出，并显示主界面。
-* **主菜单**：
-
-  * 清晰、简约的左侧垂直布局主菜单，包含"继续游戏"、"开始游戏"等核心选项。
-  * 右上角展示 LingChat Logo。
-* **全屏设置面板**：
-
-  * 通过点击主菜单的"设置"按钮触发。
-  * 以**全屏覆盖**的形式自下而上弹出，背景采用**毛玻璃效果**（`backdrop-filter`），与底层主界面内容形成视觉分离。
-  * 右上角提供一个带悬停动画的关闭按钮。
-* **导航系统**：
-
-  * **顶部横向主导航**：位于设置面板顶部，图标与文字结合，直观展示"角色"、"文本"、"高级设置"等不同分类。
-  * **滑动指示器**：当切换导航标签时，一个蓝色的指示器线段会平滑地移动到当前激活的标签下方，提供了明确的视觉反馈。
-  * **动态内容切换**：点击不同导航项会立即显示对应的内容页面，无需重新加载。
-
-## 3. 设计与资源
-
-*   **字体**: Demo 主界面使用的主要字体为 **猫啃什锦黑**，可免费商用，允许嵌入。
-    *   **字体来源**: [猫啃网](https://www.maoken.com/assorted)
-
-## 4. 重要声明
-
-- **仅为UI演示**：此 Demo 是一个独立的、用于展示视觉和交互的**样式参考**，其代码**不会被直接合并或整合**到 `LingChat` 的主项目中。
-- **未来架构待定**：`LingChat` 项目的 UI 将如何重构，以及是否会采用如 `Vue`、`React` 等现代化前端框架，**目前仍有待讨论和确认**。
-
-## 5. 如何本地运行
-
-1. 在**文件资源管理器**中，导航到 `LingChat/frontend/public/main_page_demo/` 目录。
-2. 双击 `index.html` 文件，它将在默认浏览器中打开。
-
-## 6. 运行截图
-
-![image](https://github.com/user-attachments/assets/bedcbc49-700f-4eff-90c6-e4807c8d08c4)
-
-![image](https://github.com/user-attachments/assets/72630e8c-17f8-41fd-8eb0-174aeab928ab)
-
-![image](https://github.com/user-attachments/assets/b4e8d099-c0c3-4818-90aa-83ed0eabee2d)
-
-
-
